proto: move celsig wrapper to proto pkg
