+++ pkg/proto/wrap.go
@@ -1,4 +1,4 @@
-package wrapper
+package proto
 
 import (
 	"fmt"
