+++ pkg/proto/wrap_test.go
@@ -1,0 +1,1 @@
+package proto_test
