+++ pkg/rules/celsig/library.go
@@ -1,7 +1,7 @@
 package celsig
 
 import (
-	"github.com/aquasecurity/tracee/pkg/rules/celsig/wrapper"
+	"github.com/aquasecurity/tracee/pkg/proto"
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/checker/decls"
 	"github.com/google/cel-go/common/types"
@@ -11,19 +11,19 @@ import (
 )
 
 var (
-	wrapperEvent  = decls.NewObjectType("wrapper.Event")
-	sockaddrEvent = decls.NewObjectType("wrapper.sockaddr")
+	wrapperEvent  = decls.NewObjectType("proto.Event")
+	sockaddrEvent = decls.NewObjectType("proto.sockaddr")
 )
 
 type customLib struct{}
 
 func (c *customLib) CompileOptions() []cel.EnvOption {
 	return []cel.EnvOption{
-		cel.Types(&wrapper.Event{}),
+		cel.Types(&proto.Event{}),
 		cel.Declarations(
 			decls.NewVar("input", wrapperEvent),
 			decls.NewFunction("stringArg",
-				decls.NewInstanceOverload("wrapper.Event_stringArg_string",
+				decls.NewInstanceOverload("proto.Event_stringArg_string",
 					[]*exprpb.Type{
 						wrapperEvent,
 						decls.String,
@@ -32,7 +32,7 @@ func (c *customLib) CompileOptions() []cel.EnvOption {
 				),
 			),
 			decls.NewFunction("sockaddrArg",
-				decls.NewInstanceOverload("wrapper.Event_sockaddrArg_string",
+				decls.NewInstanceOverload("proto.Event_sockaddrArg_string",
 					[]*exprpb.Type{
 						wrapperEvent,
 						decls.String,
@@ -48,11 +48,11 @@ func (c *customLib) ProgramOptions() []cel.ProgramOption {
 	return []cel.ProgramOption{
 		cel.Functions(
 			&functions.Overload{
-				Operator: "wrapper.Event_stringArg_string",
+				Operator: "proto.Event_stringArg_string",
 				Binary:   stringArg,
 			},
 			&functions.Overload{
-				Operator: "wrapper.Event_sockaddrArg_string",
+				Operator: "proto.Event_sockaddrArg_string",
 				Binary:   c.sockaddrArg,
 			},
 		),
@@ -60,17 +60,17 @@ func (c *customLib) ProgramOptions() []cel.ProgramOption {
 }
 
 func (c *customLib) sockaddrArg(lhs ref.Val, rhs ref.Val) ref.Val {
-	event := lhs.Value().(*wrapper.Event)
+	event := lhs.Value().(*proto.Event)
 	argName := rhs.Value().(string)
 	arg := argByName(event, argName)
 	if arg == nil || arg.Value.SockaddrValue == nil {
-		return types.NewObject(nil, nil, nil, &wrapper.Sockaddr{})
+		return types.NewObject(nil, nil, nil, &proto.Sockaddr{})
 	}
 	return types.NewObject(nil, nil, nil, arg.Value.SockaddrValue)
 }
 
 func stringArg(lhs ref.Val, rhs ref.Val) ref.Val {
-	event := lhs.Value().(*wrapper.Event)
+	event := lhs.Value().(*proto.Event)
 	argName := rhs.Value().(string)
 	arg := argByName(event, argName)
 	if arg == nil || arg.Value.StringValue == nil {
@@ -79,8 +79,8 @@ func stringArg(lhs ref.Val, rhs ref.Val) ref.Val {
 	return types.String(*arg.Value.StringValue)
 }
 
-// argByName is the same code as helpers.GetTraceeArgumentByName but for *wrapper.Event.
-func argByName(event *wrapper.Event, argName string) *wrapper.Argument {
+// argByName is the same code as helpers.GetTraceeArgumentByName but for *proto.Event.
+func argByName(event *proto.Event, argName string) *proto.Argument {
 	for _, arg := range event.Args {
 		if arg.Name == argName {
 			return arg
