--- pkg/rules/celsig/wrapper/wrapper_test.go
@@ -1,1 +1,0 @@
-package wrapper_test
